refactor(core, renderer): move build config files
