+++ packages/core/deprecated/component-types.ts
