+++ packages/core/deprecated/extract-schema-json.ts
@@ -1,4 +1,4 @@
-import type { ISchemaCardMessageItem, IMarkdownMessageItem } from '../../../core/src/protocols/chat';
+import type { ISchemaCardMessageItem, IMarkdownMessageItem } from '../src/protocols/chat';
 
 export function extractMarkdownAndSchemaJson(input: string): (ISchemaCardMessageItem | IMarkdownMessageItem)[] {
   const result: (ISchemaCardMessageItem | IMarkdownMessageItem)[] = [];
