+++ packages/frameworks/vue/src/package.json
