--- packages/genui-renderer/genui-sdk/core/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "@opentiny/genui-sdk",
-  "version": "0.0.1-alpha.4",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "type": "module",
-  "files": [
-    "dist"
-  ],
-  "dependencies": {
-    "zod": "^3.24.2"
-  },
-  "exports": {
-    ".": {
-      "import": "./dist/index.js",
-      "types": "./dist/index.d.ts"
-    }
-  }
-}
