+++ packages/genui-renderer/src/core/index.ts
@@ -1,6 +1,6 @@
 export * from '../../../core/src/protocols/materials';
 export * from '../../../core/src/protocols/schema';
 export * from '../../../core/src/prompt-generator/gen-prompt-config';
-export * from './component-types';
-export * from '../utils/extract-schema-json';
+export * from '../../../core/deprecated/component-types';
+export * from '../../../core/deprecated/extract-schema-json';
 export * from '../../../core/src/protocols/chat';
